+++ nav2_bt_navigator/include/nav2_bt_navigator/bt_navigator.hpp
@@ -20,15 +20,11 @@
 #include <vector>
 
 #include "geometry_msgs/msg/pose_stamped.hpp"
-#include "nav2_behavior_tree/behavior_tree_engine.hpp"
 #include "nav2_util/lifecycle_node.hpp"
 #include "nav2_msgs/action/navigate_to_pose.hpp"
 #include "nav_msgs/msg/path.hpp"
-#include "nav2_util/simple_action_server.hpp"
-#include "rclcpp_action/rclcpp_action.hpp"
-#include "tf2_ros/transform_listener.h"
-#include "tf2_ros/create_timer_ros.h"
-#include "tf2_ros/buffer.h"
+#include "nav2_bt_navigator/ros_topic_logger.hpp"
+#include "nav2_behavior_tree/bt_action_server.hpp"
 
 namespace nav2_bt_navigator
 {
@@ -37,7 +33,7 @@ namespace nav2_bt_navigator
  * @brief An action server that uses behavior tree for navigating a robot to its
  * goal position.
  */
-class BtNavigator : public nav2_util::LifecycleNode
+class BtNavigator : public nav2_behavior_tree::BtActionServer<nav2_msgs::action::NavigateToPose>
 {
 public:
   /**
@@ -50,102 +46,38 @@ public:
   ~BtNavigator();
 
 protected:
-  /**
-   * @brief Configures member variables
-   *
-   * Initializes action server for "NavigationToPose"; subscription to
-   * "goal_sub"; and builds behavior tree from xml file.
-   * @param state Reference to LifeCycle node state
-   * @return SUCCESS or FAILURE
-   */
-  nav2_util::CallbackReturn on_configure(const rclcpp_lifecycle::State & state) override;
-  /**
-   * @brief Activates action server
-   * @param state Reference to LifeCycle node state
-   * @return SUCCESS or FAILURE
-   */
-  nav2_util::CallbackReturn on_activate(const rclcpp_lifecycle::State & state) override;
-  /**
-   * @brief Deactivates action server
-   * @param state Reference to LifeCycle node state
-   * @return SUCCESS or FAILURE
-   */
-  nav2_util::CallbackReturn on_deactivate(const rclcpp_lifecycle::State & state) override;
-  /**
-   * @brief Resets member variables
-   * @param state Reference to LifeCycle node state
-   * @return SUCCESS or FAILURE
-   */
-  nav2_util::CallbackReturn on_cleanup(const rclcpp_lifecycle::State & state) override;
-  /**
-   * @brief Called when in shutdown state
-   * @param state Reference to LifeCycle node state
-   * @return SUCCESS or FAILURE
-   */
-  nav2_util::CallbackReturn on_shutdown(const rclcpp_lifecycle::State & state) override;
+  nav2_util::CallbackReturn on_configure() override;
 
-  using Action = nav2_msgs::action::NavigateToPose;
+  nav2_util::CallbackReturn on_cleanup() override;
 
-  using ActionServer = nav2_util::SimpleActionServer<Action>;
+  bool on_goal_received() override;
 
-  // Our action server implements the NavigateToPose action
-  std::unique_ptr<ActionServer> action_server_;
+  void on_loop() override;
 
-  /**
-   * @brief Action server callbacks
-   */
-  void navigateToPose();
-
-  /**
-   * @brief Goal pose initialization on the blackboard
-   */
-  void initializeGoalPose();
+  bool is_canceling() override;
 
   /**
    * @brief A subscription and callback to handle the topic-based goal published
    * from rviz
    */
   void onGoalPoseReceived(const geometry_msgs::msg::PoseStamped::SharedPtr pose);
+
+  using Action = nav2_msgs::action::NavigateToPose;
+
   rclcpp::Subscription<geometry_msgs::msg::PoseStamped>::SharedPtr goal_sub_;
-
-  /**
-   * @brief Replace current BT with another one
-   * @param bt_xml_filename The file containing the new BT
-   * @return true if the resulting BT correspond to the one in bt_xml_filename. false
-   * if something went wrong, and previous BT is mantained
-   */
-  bool loadBehaviorTree(const std::string & bt_id);
-
-  BT::Tree tree_;
-
-  // The blackboard shared by all of the nodes in the tree
-  BT::Blackboard::Ptr blackboard_;
-
-  // The XML fiñe that cointains the Behavior Tree to create
-  std::string current_bt_xml_filename_;
-  std::string default_bt_xml_filename_;
-
-  // The wrapper class for the BT functionality
-  std::unique_ptr<nav2_behavior_tree::BehaviorTreeEngine> bt_;
-
-  // Libraries to pull plugins (BT Nodes) from
-  std::vector<std::string> plugin_lib_names_;
 
   // A client that we'll use to send a command message to our own task server
   rclcpp_action::Client<Action>::SharedPtr self_client_;
-
-  // A regular, non-spinning ROS node that we can use for calls to the action client
-  rclcpp::Node::SharedPtr client_node_;
-
-  // Spinning transform that can be used by the BT nodes
-  std::shared_ptr<tf2_ros::Buffer> tf_;
-  std::shared_ptr<tf2_ros::TransformListener> tf_listener_;
 
   // Metrics for feedback
   rclcpp::Time start_time_;
   std::string robot_frame_;
   std::string global_frame_;
   double transform_tolerance_;
+
+  std::unique_ptr<RosTopicLogger> topic_logger_;
+
+  std::shared_ptr<Action::Feedback> feedback_msg_;
 };
 
 }  // namespace nav2_bt_navigator
